requirements.txt
Django~=3.0
flake8
django-ckeditor
pillow
django-bootstrap4
django-allauth
djangorestframework